# docs/api_swagger.yaml
openapi: 3.0.3
info:
  title: Intelligent Email Agent API
  description: |
    A comprehensive API for intelligent email analysis, categorization, and management.
    
    ## Features
    - **Email Processing**: Connect to IMAP servers and process emails
    - **AI Analysis**: Analyze urgency, generate summaries, and provide follow-up suggestions
    - **Categorization**: Automatically categorize emails by type and priority
    - **Notifications**: Send notifications via Slack, Telegram, and WhatsApp
    - **Voice Summaries**: Generate voice summaries using TTS
    - **Configuration Management**: Manage email settings and VIP contacts
    - **Scheduling**: Automated daily summaries and reminders
    
    ## Authentication
    This API uses email credentials for IMAP access. Ensure your email provider supports IMAP and has appropriate security settings enabled.
    
    ## Rate Limits
    - Email processing: 10 requests per minute
    - Notifications: 50 requests per minute
    - Voice generation: 20 requests per minute
    
  version: 1.0.0
  contact:
    name: Email Agent Support
    email: support@emailagent.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.emailagent.com
    description: Production server
...
paths:
...
  /api/notifications/send:
    post:
      tags:
        - Notifications
      summary: Send notification
      description: Send a notification through specified channels
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationRequest'
            example:
              message: "You have 5 new emails requiring attention"
              channels: ["slack", "telegram"]
      responses:
        '200':
          description: Notification sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
              example:
                slack: true
                telegram: true
                whatsapp: false
        '400':
          description: Invalid notification request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/notifications/daily-summary:
    post:
      tags:
        - Notifications
      summary: Send daily summary notification
      description: Send a formatted daily summary notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DailySummary'
      responses:
        '200':
          description: Daily summary notification sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /api/config/vip-contacts:
    get:
      tags:
        - Configuration
      summary: Get VIP contacts
      description: Retrieve all VIP contacts
      responses:
        '200':
          description: VIP contacts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VIPContact'
              example:
                - email: "ceo@company.com"
                  name: "CEO"
                  priority_level: "high"
                - email: "manager@company.com"
                  name: "Manager"
                  priority_level: "medium"

    post:
      tags:
        - Configuration
      summary: Add VIP contact
      description: Add a new VIP contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VIPContact'
            example:
              email: "newvip@company.com"
              name: "New VIP"
              priority_level: "high"
      responses:
        '200':
          description: VIP contact added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/config/vip-contacts/{email}:
    delete:
      tags:
        - Configuration
      summary: Delete VIP contact
      description: Remove a VIP contact
      parameters:
        - name: email
          in: path
          required: true
          description: Email address of VIP contact
          schema:
            type: string
          example: "vip@company.com"
      responses:
        '200':
          description: VIP contact deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: VIP contact not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    EmailSummary:
      type: object
      properties:
        id:
          type: string
          description: Unique email identifier
        subject:
          type: string
          description: Email subject
        sender:
          type: string
          description: Sender name
        sender_email:
          type: string
          format: email
          description: Sender email address
        received_at:
          type: string
          format: date-time
          description: Email received timestamp
        category:
          type: string
          enum: [work, personal, promotions, social, urgent, meetings, deadlines, other]
          description: Email category
        priority:
          type: string
          enum: [low, medium, high]
          description: Email priority level
        summary:
          type: string
          description: AI-generated summary
        is_read:
          type: boolean
          description: Whether email has been read
        is_replied:
          type: boolean
          description: Whether email has been replied to
        urgency_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Urgency score (0-1)
        action_required:
          type: boolean
          description: Whether action is required
        follow_up_suggestions:
          type: array
          items:
            type: string
          description: AI-generated follow-up suggestions
    DailySummary:
      type: object
      properties:
        date:
          type: string
          format: date
          description: Summary date
        total_emails:
          type: integer
          description: Total number of emails
        categories:
          type: object
          additionalProperties:
            type: integer
          description: Email count by category
        urgent_emails:
          type: array
          items:
            $ref: '#/components/schemas/EmailSummary'
          description: List of urgent emails
        unread_emails:
          type: array
          items:
            $ref: '#/components/schemas/EmailSummary'
          description: List of unread emails
        response_reminders:
          type: array
          items:
            $ref: '#/components/schemas/EmailSummary'
          description: Emails requiring response
        priority_breakdown:
          type: object
          properties:
            low:
              type: integer
            medium:
              type: integer
            high:
              type: integer
          description: Email count by priority level
    NotificationRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Notification message
        channels:
          type: array
          items:
            type: string
            enum: [slack, telegram, whatsapp]
          default: ["slack"]
          description: Notification channels
    NotificationResponse:
      type: object
      properties:
        slack:
          type: boolean
          description: Slack notification status
        telegram:
          type: boolean
          description: Telegram notification status
        whatsapp:
          type: boolean
          description: WhatsApp notification status
...
    VIPContact:
      type: object
      required:
        - email
        - name
        - priority_level
      properties:
        email:
          type: string
          format: email
          description: VIP contact email
        name:
          type: string
          description: VIP contact name
        priority_level:
          type: string
          enum: [low, medium, high]
          description: Priority level for this contact
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          description: Success message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
security:
  - ApiKeyAuth: []
...